refactor: extract describe code to k8sutils
